# docs/swagger.yaml
basePath: /
definitions:
  auth.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/auth.UserResponse'
    type: object
  auth.ErrorResponse:
    properties:
      error:
        example: invalid username or password
        type: string
    type: object
  auth.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/auth.AuthResponse'
      message:
        example: login successful
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.RegisterSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/auth.UserResponse'
      message:
        example: user registered successfully
        type: string
    type: object
  auth.UserResponse:
    properties:
      email:
        example: johndoe@example.com
        type: string
      id:
        example: 5b676171-999a-4d62-8506-c0841707b118
        type: string
      username:
        example: johndoe
        type: string
    type: object
  booking.BookingResponse:
    properties:
      created_at:
        example: "2026-09-21T11:00:00Z"
        type: string
      event_id:
        example: ae57d22b-e67e-43de-adeb-4965a6e3a334
        type: string
      id:
        example: 2c943e11-884a-4ecb-99f1-d0831707c220
        type: string
      user_id:
        example: 5b676171-999a-4d62-8506-c0841707b118
        type: string
    type: object
  booking.BookingSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/booking.BookingResponse'
      message:
        example: event booked successfully
        type: string
    type: object
  booking.ErrorResponse:
    properties:
      error:
        example: event is fully booked
        type: string
    type: object
  booking.UserBookingsSuccessResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/booking.BookingResponse'
        type: array
      message:
        example: user bookings retrieved successfully
        type: string
    type: object
  events.CreateEventRequest:
    properties:
      description:
        type: string
      ends_at:
        type: string
      location:
        minLength: 3
        type: string
      quota:
        minimum: 1
        type: integer
      starts_at:
        type: string
      title:
        minLength: 3
        type: string
    required:
    - ends_at
    - location
    - quota
    - starts_at
    - title
    type: object
  events.ErrorResponse:
    properties:
      error:
        example: event not found
        type: string
    type: object
  events.EventResponse:
    properties:
      created_at:
        example: "2026-09-21T09:31:17Z"
        type: string
      description:
        example: Deep dive into goroutines, channels, and row-level locking.
        type: string
      ends_at:
        example: "2026-10-01T17:00:00Z"
        type: string
      id:
        example: ae57d22b-e67e-43de-adeb-4965a6e3a334
        type: string
      location:
        example: Jakarta Convention Center
        type: string
      quota:
        example: 100
        type: integer
      starts_at:
        example: "2026-10-01T09:00:00Z"
        type: string
      title:
        example: Go Concurrency & Backend Masterclass
        type: string
    type: object
  events.EventSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/events.EventResponse'
      message:
        example: event retrieved successfully
        type: string
    type: object
  events.EventsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/events.EventResponse'
        type: array
      message:
        example: events retrieved successfully
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: High-performance Event Management & Concurrency-Safe Ticket Booking
    REST API in Go.
  title: Event Booking REST API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return signed JWT bearer token
      parameters:
      - description: User Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with JWT token
          schema:
            $ref: '#/definitions/auth.LoginSuccessResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with username, email, and password
      parameters:
      - description: User Registration Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/auth.RegisterSuccessResponse'
        "400":
          description: Validation error or username/email taken
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /events:
    get:
      description: Retrieve a list of all available events
      produces:
      - application/json
      responses:
        "200":
          description: List of events
          schema:
            $ref: '#/definitions/events.EventsListResponse'
        "400":
          description: Database error
          schema:
            $ref: '#/definitions/events.ErrorResponse'
      summary: Get all events
      tags:
      - Events
  /events/{id}:
    get:
      description: Retrieve detailed information for a single event
      parameters:
      - description: Event UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Event details
          schema:
            $ref: '#/definitions/events.EventSuccessResponse'
        "404":
          description: Event not found
          schema:
            $ref: '#/definitions/events.ErrorResponse'
      summary: Get event by ID
      tags:
      - Events
  /events/{id}/book:
    post:
      description: Reserve a ticket for an event with row-level lock concurrency safety
        (Requires JWT Authentication)
      parameters:
      - description: Event UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Event booked successfully
          schema:
            $ref: '#/definitions/booking.BookingSuccessResponse'
        "400":
          description: Event fully booked or already booked by user
          schema:
            $ref: '#/definitions/booking.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/booking.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Book an event ticket
      tags:
      - Bookings
  /events/create:
    post:
      consumes:
      - application/json
      description: Create a new event with schedule and seat quota (Requires JWT Authentication)
      parameters:
      - description: Event Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/events.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Event created successfully
          schema:
            $ref: '#/definitions/events.EventSuccessResponse'
        "400":
          description: Validation error or invalid schedule
          schema:
            $ref: '#/definitions/events.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/events.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new event
      tags:
      - Events
  /me/bookings:
    get:
      description: Retrieve list of all tickets booked by the currently authenticated
        user
      produces:
      - application/json
      responses:
        "200":
          description: User booking history
          schema:
            $ref: '#/definitions/booking.UserBookingsSuccessResponse'
        "400":
          description: Database error
          schema:
            $ref: '#/definitions/booking.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/booking.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user booking history
      tags:
      - Bookings
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"